pyproject.toml
[project]
name = "boosters-tt"
version = "0.1.0"
description = "Technical task for Boosters - Scrapper ETL workflow"
authors = [
    {name = "vvk",email = "koval.vladislavv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "scrapy (>=2.13.3,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "polars (>=1.31.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "apache-airflow[postgres] (==2.9.1)",
    "plotly (>=6.2.0,<7.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "pandas (>=2.3.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "arxiv_scrapper" },
    { include = "etl" }
]